refactor nostr modules
